BackendPart/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.30.0
asyncpg==0.30.0
python-dotenv==1.2.1
yoyo-migrations==8.2.0
python-multipart==0.0.9
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.12.4
pydantic_core==2.41.5
typing_extensions==4.15.0
psycopg2-binary==2.9.10
httpx